requirements.txt
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.18
python-dotenv>=1.0.0
streamlit>=1.28.0
PyPDF2>=3.0.1
tiktoken>=0.5.2 